requirements.txt
Django
Pillow
django-widget-tweaks
django-crispy-forms
python-decouple
six
django-gravatar2
django-relativedelta
django-model-utils
swapper
djangorestframework